♻️  zb: Move InterfaceRef & InterfaceDeref* in interface mod
