+++ zbus/src/object_server/mod.rs
@@ -6,7 +6,6 @@ use std::{
     collections::{hash_map::Entry, HashMap},
     fmt::Write,
     marker::PhantomData,
-    ops::{Deref, DerefMut},
     sync::Arc,
 };
 use tracing::{debug, instrument, trace, trace_span, Instrument};
@@ -16,7 +15,7 @@ use zbus_names::InterfaceName;
 use zvariant::{ObjectPath, OwnedObjectPath, OwnedValue, Signature, Type, Value};
 
 use crate::{
-    async_lock::{RwLock, RwLockReadGuard, RwLockWriteGuard},
+    async_lock::RwLock,
     connection::WeakConnection,
     fdo,
     fdo::{Introspectable, ManagedObjects, ObjectManager, Peer, Properties},
@@ -26,160 +25,10 @@ use crate::{
 
 mod interface;
 pub(crate) use interface::ArcInterface;
-pub use interface::{DispatchResult, Interface};
+pub use interface::{DispatchResult, Interface, InterfaceDeref, InterfaceDerefMut, InterfaceRef};
 
 mod signal_context;
 pub use signal_context::SignalContext;
-
-/// Opaque structure that derefs to an `Interface` type.
-pub struct InterfaceDeref<'d, I> {
-    iface: RwLockReadGuard<'d, dyn Interface>,
-    phantom: PhantomData<I>,
-}
-
-impl<I> Deref for InterfaceDeref<'_, I>
-where
-    I: Interface,
-{
-    type Target = I;
-
-    fn deref(&self) -> &I {
-        self.iface.downcast_ref::<I>().unwrap()
-    }
-}
-
-/// Opaque structure that mutably derefs to an `Interface` type.
-pub struct InterfaceDerefMut<'d, I> {
-    iface: RwLockWriteGuard<'d, dyn Interface>,
-    phantom: PhantomData<I>,
-}
-
-impl<I> Deref for InterfaceDerefMut<'_, I>
-where
-    I: Interface,
-{
-    type Target = I;
-
-    fn deref(&self) -> &I {
-        self.iface.downcast_ref::<I>().unwrap()
-    }
-}
-
-impl<I> DerefMut for InterfaceDerefMut<'_, I>
-where
-    I: Interface,
-{
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.iface.downcast_mut::<I>().unwrap()
-    }
-}
-
-/// Wrapper over an interface, along with its corresponding `SignalContext`
-/// instance. A reference to the underlying interface may be obtained via
-/// [`InterfaceRef::get`] and [`InterfaceRef::get_mut`].
-pub struct InterfaceRef<I> {
-    ctxt: SignalContext<'static>,
-    lock: Arc<RwLock<dyn Interface>>,
-    phantom: PhantomData<I>,
-}
-
-impl<I> InterfaceRef<I>
-where
-    I: 'static,
-{
-    /// Get a reference to the underlying interface.
-    ///
-    /// **WARNING:** If methods (e.g property setters) in `ObjectServer` require `&mut self`
-    /// `ObjectServer` will not be able to access the interface in question until all references
-    /// of this method are dropped; it is highly recommended that the scope of the interface
-    /// returned is restricted.
-    pub async fn get(&self) -> InterfaceDeref<'_, I> {
-        let iface = self.lock.read().await;
-
-        iface
-            .downcast_ref::<I>()
-            .expect("Unexpected interface type");
-
-        InterfaceDeref {
-            iface,
-            phantom: PhantomData,
-        }
-    }
-
-    /// Get a reference to the underlying interface.
-    ///
-    /// **WARNINGS:** Since the `ObjectServer` will not be able to access the interface in question
-    /// until the return value of this method is dropped, it is highly recommended that the scope
-    /// of the interface returned is restricted.
-    ///
-    /// # Errors
-    ///
-    /// If the interface at this instance's path is not valid, an `Error::InterfaceNotFound` error
-    /// is returned.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use std::error::Error;
-    /// # use async_io::block_on;
-    /// # use zbus::{Connection, interface};
-    ///
-    /// struct MyIface(u32);
-    ///
-    /// #[interface(name = "org.myiface.MyIface")]
-    /// impl MyIface {
-    ///    #[zbus(property)]
-    ///    async fn count(&self) -> u32 {
-    ///        self.0
-    ///    }
-    /// }
-    ///
-    /// # block_on(async {
-    /// // Set up connection and object_server etc here and then in another part of the code:
-    /// # let connection = Connection::session().await?;
-    /// #
-    /// # let path = "/org/zbus/path";
-    /// # connection.object_server().at(path, MyIface(22)).await?;
-    /// let object_server = connection.object_server();
-    /// let iface_ref = object_server.interface::<_, MyIface>(path).await?;
-    /// let mut iface = iface_ref.get_mut().await;
-    /// iface.0 = 42;
-    /// iface.count_changed(iface_ref.signal_context()).await?;
-    /// # Ok::<_, Box<dyn Error + Send + Sync>>(())
-    /// # })?;
-    /// #
-    /// # Ok::<_, Box<dyn Error + Send + Sync>>(())
-    /// ```
-    pub async fn get_mut(&self) -> InterfaceDerefMut<'_, I> {
-        let mut iface = self.lock.write().await;
-
-        iface
-            .downcast_ref::<I>()
-            .expect("Unexpected interface type");
-        iface
-            .downcast_mut::<I>()
-            .expect("Unexpected interface type");
-
-        InterfaceDerefMut {
-            iface,
-            phantom: PhantomData,
-        }
-    }
-
-    pub fn signal_context(&self) -> &SignalContext<'static> {
-        &self.ctxt
-    }
-}
-
-impl<I> Clone for InterfaceRef<I> {
-    fn clone(&self) -> Self {
-        Self {
-            ctxt: self.ctxt.clone(),
-            lock: self.lock.clone(),
-            phantom: PhantomData,
-        }
-    }
-}
 
 #[derive(Default, Debug)]
 pub(crate) struct Node {
